+++ .github/workflows/on-pull-request.yml
@@ -36,7 +36,6 @@ jobs:
         npm run lint
     
     - name: Build frontend
-      if: false
       run: |
         cd frontend 
         npm run build --if-present
