+++ frontend/src/pages/log_history/LogHistory.css
