+++ frontend/src/pages/log_history/__tests__/LogHistory.test.js
@@ -1,0 +1,3 @@
+test("empty test", () => {
+  expect(true).toBe(true)
+})
