+++ frontend/src/pages/manual_entry/ManualEntry.jsx
