+++ frontend/src/pages/manual_entry/__tests__/ManualEntry.test.js
@@ -1,0 +1,3 @@
+test("empty test", () => {
+  expect(true).toBe(true)
+})
