+++ frontend/src/pages/new_log/NewLog.css
