+++ frontend/src/pages/new_log/__tests__/NewLog.test.js
@@ -1,0 +1,3 @@
+test("empty test", () => {
+  expect(true).toBe(true)
+})
