+++ frontend/src/pages/upload_photo/UploadPhoto.css
