+++ frontend/src/pages/upload_photo/__tests__/UploadPhoto.test.js
@@ -1,0 +1,3 @@
+test("empty test", () => {
+  expect(true).toBe(true)
+})
